feat(molecules): move Notification compononent to molecules folder, update imports
